requirements.txt
aiofiles==0.8.0
aiohttp==3.8.1
fastapi==0.70.0
opencv-python-headless==4.5.4.60
Pillow==8.4.0
python-multipart==0.0.5
tensorflow==2.7.0
tensorflow-hub==0.12.0
uvicorn==0.15.0
filelock==3.4.0
python-dotenv==0.19.2